requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
cloudinary==1.28.1
colorthief==0.2.1
cycler==0.11.0
fonttools==4.28.5
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
toml==0.10.2
urllib3==1.26.8
